code cleanup to remove warnings

+++ src-dut/delay_calculator.cpp
@@ -189,7 +189,7 @@ int compute_delays(dctk::CellLib* cell_lib, dctk::CircuitPtrVec* circuitMgr, spe
 
 
     // Example on how to set return values for delay and slew
-    int numCircuits = circuitMgr->size();
+    size_t numCircuits = circuitMgr->size();
     for (size_t i = 0; i < numCircuits; i++) {
         compute_delays_bilinear(cell_lib, (*circuitMgr)[i], spef);
     }
